Hide keyboard on start for the default transitions
The most common use case when switching Scenes is to
hide the keyboard. If requirements differ for specific
Scenes, the transition can be overridden for them using
the TransitionFactory.
