+++ samples/notes-app/bravo/android/src/main/java/com/nhaarman/bravo/notesapp/android/TransitionFactoryProvider.kt
@@ -1,11 +1,23 @@
 package com.nhaarman.bravo.notesapp.android
 
-import com.nhaarman.bravo.android.transition.DefaultTransitionFactory
+import com.nhaarman.bravo.android.transition.FadeOutToBottomTransition
 import com.nhaarman.bravo.android.transition.TransitionFactory
+import com.nhaarman.bravo.android.transition.hideKeyboardOnStart
+import com.nhaarman.bravo.android.transition.transitionFactory
+import com.nhaarman.bravo.notesapp.presentation.createitem.CreateItemScene
+import com.nhaarman.bravo.notesapp.presentation.edititem.EditItemScene
+import com.nhaarman.bravo.notesapp.presentation.itemlist.ItemListScene
 
 object TransitionFactoryProvider {
 
     val transitionFactory: TransitionFactory by lazy {
-        DefaultTransitionFactory(ViewFactoryProvider.viewFactory)
+        transitionFactory(ViewFactoryProvider.viewFactory) {
+            val fadeOutToBottom =
+                FadeOutToBottomTransition.from(ViewFactoryProvider.viewFactory)
+                    .hideKeyboardOnStart()
+
+            (CreateItemScene::class.java to ItemListScene::class.java) use fadeOutToBottom
+            (EditItemScene::class.java to ItemListScene::class.java) use fadeOutToBottom
+        }
     }
 }
